Using WebEncoders' Base64Url encode/decode functionality
